web: moving the tests up a level
