+++ azurerm/internal/services/web/resource_arm_app_service_active_slot_test.go
@@ -1,4 +1,4 @@
-package tests
+package web_test
 
 import (
 	"fmt"
