+++ azurerm/internal/services/web/resource_arm_app_service_custom_hostname_binding_test.go
@@ -1,4 +1,4 @@
-package tests
+package web_test
 
 import (
 	"fmt"
