+++ azurerm/internal/services/web/resource_arm_function_app_slot_virtual_network_association_test.go
@@ -1,4 +1,4 @@
-package tests
+package web_test
 
 import (
 	"fmt"
